Redesigned ackview, small fixes to aboutview
