+++ CodeEdit/Features/About/Views/AboutView.swift
@@ -33,8 +33,10 @@ public struct AboutView: View {
                 bottomMetaData
                 actionButtons
             }
-            .padding([.trailing, .bottom])
+            .padding([.trailing, .vertical])
         }
+        .background(.regularMaterial)
+        .edgesIgnoringSafeArea(.top)
     }
 
     // MARK: Sub-Views
@@ -73,16 +75,15 @@ public struct AboutView: View {
             Button {
                 AcknowledgementsView().showWindow(width: 300, height: 400)
             } label: {
-                Text("Acknowledgments")
-                    .frame(maxWidth: .infinity)
+                Text("Acknowledgements")
+                    .foregroundColor(.primary)
             }
+
             Button {
-                guard let url = URL(string: "https://github.com/CodeEditApp/CodeEdit/blob/main/LICENSE.md")
-                else { return }
-                openURL(url)
+                openURL(URL(string: "https://github.com/CodeEditApp/CodeEdit/blob/main/LICENSE.md")!)
             } label: {
                 Text("License Agreement")
-                    .frame(maxWidth: .infinity)
+                    .foregroundColor(.primary)
             }
         }
     }
@@ -93,53 +94,5 @@ public struct AboutView: View {
             size: NSSize(width: width, height: height)
         )
         .showWindow(nil)
-    }
-}
-
-final class AboutViewWindowController: NSWindowController {
-    convenience init<T: View>(view: T, size: NSSize) {
-        let hostingController = NSHostingController(rootView: view)
-        // New window holding our SwiftUI view
-        let window = NSWindow(contentViewController: hostingController)
-        self.init(window: window)
-        window.setContentSize(size)
-        window.styleMask.remove(.resizable)
-        window.styleMask.insert(.fullSizeContentView)
-        window.alphaValue = 0.5
-        window.styleMask.remove(.miniaturizable)
-    }
-
-    override func showWindow(_ sender: Any?) {
-        window?.center()
-        window?.alphaValue = 0.0
-
-        super.showWindow(sender)
-
-        window?.animator().alphaValue = 1.0
-
-        // close the window when the escape key is pressed
-        NSEvent.addLocalMonitorForEvents(matching: .keyDown) { event in
-            guard event.keyCode == 53 else { return event }
-
-            self.closeAnimated()
-
-            return nil
-        }
-
-        window?.collectionBehavior = [.transient, .ignoresCycle]
-        window?.isMovableByWindowBackground = true
-        window?.titlebarAppearsTransparent = true
-        window?.titleVisibility = .hidden
-        window?.isExcludedFromWindowsMenu = true
-    }
-
-    func closeAnimated() {
-        NSAnimationContext.beginGrouping()
-        NSAnimationContext.current.duration = 0.4
-        NSAnimationContext.current.completionHandler = {
-            self.close()
-        }
-        window?.animator().alphaValue = 0.0
-        NSAnimationContext.endGrouping()
     }
 }
